[Feat/90] 멀티플레이어 거리 계산 및 점수 부여 기능 서버 위임 및 Coordinate 구조 리팩토링

+++ src/main/java/com/kospot/domain/multi/gamePlayer/entity/GamePlayer.java
@@ -38,7 +38,7 @@ public class GamePlayer {
     private Integer teamNumber;
 
     private Integer roundRank; // 해당 라운드 순위
-    private int totalScore;
+    private double totalScore;
 
     @Enumerated(EnumType.STRING)
     private GamePlayerStatus status;
@@ -87,7 +87,7 @@ public class GamePlayer {
         this.roundRank = rank;
     }
 
-    public void addScore(int points) {
+    public void addScore(double points) {
         this.totalScore += points;
     }
 }
